+++ lib/hive/adapter.dart
@@ -5,8 +5,8 @@ import 'package:mimir/credential/entity/login_status.dart';
 import 'package:mimir/entity/campus.dart';
 import 'package:mimir/life/electricity/entity/balance.dart';
 import 'package:mimir/life/expense_records/entity/local.dart';
+import 'package:mimir/school/ywb/entity/meta.dart';
 import 'package:mimir/school/ywb/entity/application.dart';
-import 'package:mimir/school/ywb/entity/message.dart';
 import 'package:mimir/school/exam_arrange/entity/exam.dart';
 import 'package:mimir/school/oa_announce/entity/page.dart';
 import 'package:mimir/school/exam_result/entity/result.dart';
